Got rid of an unnecessary directory; moved the tests up one level in the directory structure.
