+++ integration-tests/src/__tests__/containers/data/FooEmitter.js
