+++ integration-tests/src/__tests__/containers/pages/AttributesOnRootContainer.js
