+++ integration-tests/src/__tests__/containers/pages/AttributesOnRootElement.js
