+++ integration-tests/src/__tests__/containers/pages/PropsFromEmitterRootElement.js
