+++ integration-tests/src/__tests__/containers/pages/SinglePromiseInArray.js
