+++ integration-tests/src/__tests__/containers/pages/SinglePromiseNoArray.js
