+++ integration-tests/src/__tests__/containers/pages/SingleRootElementInArray.js
