+++ integration-tests/src/__tests__/containers/pages/SingleRootElementInRootContainer.js
