+++ integration-tests/src/__tests__/internalServerError/InternalServerErrorSpec.js
@@ -1,4 +1,4 @@
-var helper = require("../../../specRuntime/testHelper");
+var helper = require("../../specRuntime/testHelper");
 var Browser = require("zombie");
 
 describe("A 500 internal server error page", () => {
