+++ integration-tests/src/__tests__/redirectForward/PermanentRedirectWithDocumentPage.js
