+++ integration-tests/src/__tests__/simpleRender/GoodbyeWorldPage.js
