+++ integration-tests/src/__tests__/simpleRender/HelloWorldSpec.js
@@ -1,4 +1,4 @@
-var helper = require("../../../specRuntime/testHelper");
+var helper = require("../../specRuntime/testHelper");
 
 describe("A basic page", () => {
